Don't use CPU/GPU when opacity zero.
If spinning is used, but opacity is zero, disable the animation via coercion of the Spin dependency property.

To test this, run the WPF example app and observe the GPU column in Process Explorer (from Sysinternals).
With spinning enabled, slide the opacity to zero and notice how the animation stops.